Updated seeds to not import supertest dev package
