+++ helpers/tests.js
@@ -3,29 +3,8 @@ const request = require('supertest');
 const mongoose = require('mongoose');
 const server = require('../api/server');
 const connect = require('../api/database');
+const { dropAllCollections, removeAllCollections } = require('./helpers');
 
-async function removeAllCollections() {
-  const collections = Object.keys(mongoose.connection.collections);
-  await collections.map((colname) => mongoose.connection.collections[colname])
-    .forEach((col) => col.deleteMany());
-}
-
-async function dropAllCollections() {
-  const collections = Object.keys(mongoose.connection.collections);
-  await collections.map((colname) => {
-    const collection = mongoose.connection.collections[colname];
-    try {
-      return collection.drop();
-    } catch (error) {
-      // Sometimes this error happens, but you can safely ignore it
-      if (error.message === 'ns not found') return false;
-      // This error occurs when you use it.todo. You can
-      // safely ignore this error too
-      if (error.message.includes('a background operation is currently running')) return false;
-      throw new Error(error.message);
-    }
-  });
-}
 
 module.exports = {
   initTest: () => {
